+++ stream-chat-android-offline/src/main/java/io/getstream/chat/android/offline/ChatDomain.kt
@@ -350,6 +350,14 @@ public sealed interface ChatDomain {
      * @see io.getstream.chat.android.offline.utils.RetryPolicy
      */
     @CheckResult
+    @Deprecated(
+        message = "cancelMessage is deprecated. Use extension function ChatClient::cancelMessage instead",
+        replaceWith = ReplaceWith(
+            expression = "ChatClient.instance().cancelMessage(message)",
+            imports = arrayOf("io.getstream.chat.android.client.ChatClient")
+        ),
+        level = DeprecationLevel.WARNING
+    )
     public fun cancelMessage(message: Message): Call<Boolean>
 
     /**